macOS 10.13.2 SDK @ Xcode 9.2
https://download.developer.apple.com/Developer_Tools/Xcode_9.2/Xcode_9.2.xip

+++ MacOSX10.13.sdk/System/Library/Frameworks/Kernel.framework/Versions/A/Headers/i386/proc_reg.h
@@ -304,9 +304,9 @@ static inline uintptr_t get_cr3_base(void)
 	return(cr3 & ~(0xFFFULL));
 }
 
-static inline void set_cr3_composed(uintptr_t base, uint16_t pcid, uint32_t preserve)
+static inline void set_cr3_composed(uintptr_t base, uint16_t pcid, uint64_t preserve)
 {
-	__asm__ volatile("mov %0, %%cr3" : : "r" (base | pcid | ( ( (uint64_t)preserve) << 63) ) );
+	__asm__ volatile("mov %0, %%cr3" : : "r" (base | pcid | ( (preserve) << 63) ) );
 }
 
 static inline uintptr_t get_cr4(void)
