+++ MacOSX10.13.sdk/System/Library/Frameworks/Quartz.framework/Versions/A/Frameworks/PDFKit.framework/Versions/A/Headers/PDFAnnotationTextWidget.h
@@ -32,7 +32,7 @@ PDFKIT_CLASS_DEPRECATED(10_4, 10_12, NA, NA)
 
 - (PDFKitPlatformColor *) backgroundColor;
 - (void) setBackgroundColor: (PDFKitPlatformColor *) color;
-- (int) rotation;
+- (NSInteger) rotation;
 - (void) setRotation: (int) rotation;
 
 // Font characteristics associated with the text field.
